# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "api/credential/employment": {
            "post": {
                "tags": [
                    "Credential"
                ],
                "summary": "user update employment credentials",
                "description": "user update employment credentials",
                "operationId": "11aac243f3e8f8a1b728c9314a9ce382",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "position",
                                    "company",
                                    "start_year",
                                    "end_year"
                                ],
                                "properties": {
                                    "position": {
                                        "type": "nullable|string|max:50"
                                    },
                                    "company": {
                                        "type": "nullable|string"
                                    },
                                    "start_year": {
                                        "type": "nullable|integer"
                                    },
                                    "end_year": {
                                        "type": "nullable|integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/credential/employment/{id}": {
            "post": {
                "tags": [
                    "Credential"
                ],
                "summary": "user update employment credentials",
                "description": "user update employment credentials",
                "operationId": "c29060fb24aa1309052dc66dc592b0e1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "position",
                                    "company",
                                    "start_year",
                                    "end_year"
                                ],
                                "properties": {
                                    "position": {
                                        "type": "nullable|string|max:50"
                                    },
                                    "company": {
                                        "type": "nullable|string"
                                    },
                                    "start_year": {
                                        "type": "nullable|integer"
                                    },
                                    "end_year": {
                                        "type": "nullable|integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/credential/employment/{user_id}": {
            "get": {
                "tags": [
                    "Credential"
                ],
                "summary": "Get user employment details",
                "description": "Get user employment details",
                "operationId": "a7055242607b6c91e71c6bb57073c0a7",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/deleteEmployment/{id}": {
            "delete": {
                "tags": [
                    "Delete"
                ],
                "summary": "Delete data by id",
                "description": "Delete data by id",
                "operationId": "a394e193bb482807528773d70defec8a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "type": "integer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}